requirements.txt
pandas
numpy
sqlalchemy
psycopg2-binary
python-dotenv
scikit-learn
jupyter
catboost